requirements.txt
paho-mqtt
python-dotenv
flet
requests
pydantic
aiohttp